+++ packages/plugin-inquirer-cli/test/src/Cli.spec.js
@@ -671,7 +671,7 @@ describe("Cli", () => {
 
     it("should print current mock in yellow if there are custom routes variants", async () => {
       coreInstance.mock.collections.selected = "foo";
-      coreInstance.mock.customRoutesVariants = ["foo-variant", "foo-variant-2"];
+      coreInstance.mock.customRouteVariants = ["foo-variant", "foo-variant-2"];
       await cli.start();
       expect(cli._header()[2]).toEqual(
         expect.stringContaining(chalk.yellow("foo (custom variants: foo-variant,foo-variant-2)"))