+++ packages/plugin-admin-api/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat: Use core.mock.restoreRouteVariants
 - feat: Use core.mock.routes.plain getter
 - feat: Use core.mock.routes.plainVariants getter
+- feat: Use core.mock.customRouteVariants getter
 
 ## [3.1.0] - 2022-06-27